streamlit_app/grok31/requirements.txt
#env variables
python-dotenv

#sdk
openai

# streamlit 
streamlit